double values transfromed to float so we have less casting in android, planet, spaceship and background texture index is now a server variable (so every one seeing the same map, still random the whole thing)

+++ src/net/neczpal/interstellarwar/desktop/geom/Arrow.java
@@ -4,14 +4,14 @@ package net.neczpal.interstellarwar.desktop.geom;
 public class Arrow extends Line {
 	private Triangle triangle;
 
-	public Arrow (Point from, Point to, double arrowSize) {
+	public Arrow (Point from, Point to, float arrowSize) {
 		super (from, to);
 
-		double length = from.distance (to);
-		double lx = to.getX () - from.getX ();
-		double ly = to.getY () - from.getY ();
-		double hx = lx / length * arrowSize;
-		double hy = ly / length * arrowSize;
+		float length = from.distance (to);
+		float lx = to.getX () - from.getX ();
+		float ly = to.getY () - from.getY ();
+		float hx = lx / length * arrowSize;
+		float hy = ly / length * arrowSize;
 
 		Point a = new Point (to.getX (), to.getY ());
 		Point b = new Point (to.getX () - hx / 2, to.getY () - hy / 2);
