+++ src/net/neczpal/interstellarwar/desktop/geom/Quad.java
@@ -40,14 +40,14 @@ public class Quad {
 		glEnd ();
 	}
 
-	public void move (double dx, double dy) {
+	public void move (float dx, float dy) {
 		a.move (dx, dy);
 		b.move (dx, dy);
 		c.move (dx, dy);
 		d.move (dx, dy);
 	}
 
-	public void rotate (Point p, double angle) {
+	public void rotate (Point p, float angle) {
 		a.rotate (p, angle);
 		b.rotate (p, angle);
 		c.rotate (p, angle);
